+++ testing_migrations/phoenix/20160921152624_add_data_after_changes.php
@@ -2,7 +2,7 @@
 
 namespace Phoenix\TestingMigrations;
 
-use Nette\Utils\DateTime;
+use DateTime;
 use Phoenix\Migration\AbstractMigration;
 
 class AddDataAfterChanges extends AbstractMigration