+++ CHANGELOG.md
@@ -2,8 +2,19 @@
 
 ### [Unreleased][unreleased]
 
+#### Changed
+- Replaced nette/finder with symfony/finder
+
+#### Removed
+- Removed nette/utils
+
+### [0.13.0] - 2018-03-14
+#### Changed
+- second parameter of methods fetchAll and fetch `string $fields = '*'` has been changed to `array $fields = ['*']` (BC Break - fix: change string to array)
+
 #### Added
-- Added methods tableExist, tableColumnExists and getTable to migration
+- methods tableExists, tableColumnExists and getTable to migration
+- posibility to add an autoincrement primary key on an existing table
 
 #### Fixed
 - Add a primary key on an existing table
@@ -160,7 +171,8 @@
 - supported methods in migrations: addColumn, addIndex, addForeignKey, dropColumn, dropIndex, dropForeignKey
 - supported column types: string, integer, boolean, text, datetime, uuid, json, char
 
-[unreleased]: https://github.com/lulco/phoenix/compare/0.12.1...HEAD
+[unreleased]: https://github.com/lulco/phoenix/compare/0.13.0...HEAD
+[0.13.0]: https://github.com/lulco/phoenix/compare/0.12.1...0.13.0
 [0.12.1]: https://github.com/lulco/phoenix/compare/0.12.0...0.12.1
 [0.12.0]: https://github.com/lulco/phoenix/compare/0.11.0...0.12.0
 [0.11.0]: https://github.com/lulco/phoenix/compare/0.10.0...0.11.0